brick: Add a nic brick
This brick is used to communicate with a ethernet, or a pcap port
The reading patrt is implemented in the pull callback and the writing
part is in the burst callback.

This brick is configure by the port id of the ethernet to use and by the
output side to use.

Signed-off-by: Matthias Gatto <matthias.gatto@outscale.com>