api/server: fix security group rule checking
When a TCP/UDP rule is added, a port range must be defined.
Port range start must not be greater that port range end.
This is not currently a bug as firewall rules building in graph.c
detects this case but API_0::validate_sg_rule() must have checked it
before.

Signed-off-by: Jerome Jutteau <jerome.jutteau@outscale.com>