tests: remove set -e in function.sh
remove set -e/+e, but add a function "fail".

This basically print a error message, clean_all and exit 1, this allow
to controlle the error flow of the tests, and remove ~50 lines of
duplicated code.

The controlle of the error flow is a big plus, as now the tests can fail
only when we think they can fail, so they tests only what we want them to
tests.

remove some trailling-whitespace too

Signed-off-by: Matthias Gatto <matthias.gatto@outscale.com>